+++ libshvhistory/rpchandler_history.c
@@ -7,9 +7,14 @@
 #define RECORDS_PREFIX ".records/"
 #define FILES_PREFIX ".files/"
 
+#define obstack_chunk_alloc malloc
+#define obstack_chunk_free free
+
 struct rpchandler_history {
 	bool has_records;
 	bool has_files;
+	bool has_getlog;
+	const char **signals;
 	struct rpchandler_history_facilities *facilities;
 };
 
@@ -20,6 +25,14 @@ static void rpc_ls(void *cookie, struct rpchandler_ls *ctx) {
 			rpchandler_ls_result(ctx, ".records");
 		if (history->has_files)
 			rpchandler_ls_result(ctx, ".files");
+		if (history->has_getlog) {
+			/* Handle virtual tree generated for getLog method. This takes care
+			 * of the root nodes.
+			 */
+			for (const char **s = history->signals; s && *s; s++)
+				rpchandler_ls_result_fmt(
+					ctx, "%.*s", (int)(strchrnul(*s, '/') - *s), *s);
+		}
 	} else if (!strcmp(ctx->path, ".records")) {
 		for (struct rpchandler_history_records **record =
 				 history->facilities->records;
@@ -29,6 +42,25 @@ static void rpc_ls(void *cookie, struct rpchandler_ls *ctx) {
 		for (struct rpchandler_history_files **file = history->facilities->files;
 			 file && *file; file++)
 			rpchandler_ls_result(ctx, (*file)->name);
+	} else if (history->has_getlog) {
+		/* Handle getLog virtual tree for all subnodes. This requires parsing
+		 * of signal paths to separate nodes and correctly identifying children
+		 * nodes present on the given path.
+		 */
+		for (const char **s = history->signals; s && *s; s++) {
+			/* Continue if ctx->path substring does not start at given signal
+			 * path
+			 */
+			size_t len = strlen(ctx->path);
+			if (strncmp(*s, ctx->path, len) || (*s)[len] != '/')
+				continue;
+
+			/* Get the starting position of path's children. */
+			const char *start = *s + len;
+			/* The children's end is the next '/' character. */
+			const char *end = strchrnul(++start, '/');
+			rpchandler_ls_result_fmt(ctx, "%.*s", (int)(end - start), start);
+		}
 	}
 }
 
@@ -50,6 +82,19 @@ static void rpc_dir(void *cookie, struct rpchandler_dir *ctx) {
 		for (struct rpchandler_history_files **file = history->facilities->files;
 			 file && *file; file++) {
 			// TODO: files based logging
+		}
+	} else if (history->has_getlog) {
+		for (const char **s = history->signals; s && *s; s++) {
+			size_t len = strlen(ctx->path);
+			if ((ctx->path[0] == '\0') ||
+				(!strncmp(*s, ctx->path, len) &&
+					((*s)[len] == '\0' || (*s)[len] == '/'))) {
+				/* add getLog method if *s starts with ctx->path and is either
+				 * entire ctx->path or whole path until '/' character.
+				 */
+				rpchandler_dir_result(ctx, &rpchistory_getlog);
+				break;
+			}
 		}
 	}
 }
@@ -128,24 +173,62 @@ static enum rpchandler_msg_res handle_files_history(void) {
 	return RPCHANDLER_MSG_SKIP;
 }
 
+static void rpcgetlog_result(struct rpchandler_history_facilities *facilities,
+	struct rpchistory_getlog_request *request, struct rpchandler_msg *ctx,
+	const char *path) {
+	cp_pack_t pack = rpchandler_msg_new_response(ctx);
+	cp_pack_list_begin(pack);
+	if (facilities->pack_getlog(
+			facilities, request, pack, rpchandler_obstack(ctx), path)) {
+		cp_pack_container_end(pack);
+		rpchandler_msg_send_response(ctx, pack);
+	} else {
+		rpchandler_msg_drop(ctx);
+		rpchandler_msg_send_error(ctx, RPCERR_INTERNAL_ERR,
+			"Time in the logs is not valid, not possible to provide records.");
+	}
+}
+
 static enum rpchandler_msg_res rpc_msg(void *cookie, struct rpchandler_msg *ctx) {
 	struct rpchandler_history *history = cookie;
-	if (ctx->meta.type == RPCMSG_T_REQUEST) {
-		if (history->has_records &&
-			!strncmp(ctx->meta.path, RECORDS_PREFIX, strlen(RECORDS_PREFIX))) {
-			for (struct rpchandler_history_records **record =
-					 history->facilities->records;
-				 record && *record; record++) {
-				if (!strcmp(ctx->meta.path + strlen(RECORDS_PREFIX), (*record)->name))
-					return handle_records_history(ctx, *record);
-			}
-		} else if (history->has_files &&
-			!strncmp(ctx->meta.path, FILES_PREFIX, strlen(FILES_PREFIX))) {
-			for (struct rpchandler_history_files **file = history->facilities->files;
-				 file && *file; file++) {
-				if (!strcmp(ctx->meta.path + strlen(FILES_PREFIX), (*file)->name))
-					return handle_files_history();
-			}
+	if (ctx->meta.type != RPCMSG_T_REQUEST)
+		return RPCHANDLER_MSG_SKIP;
+
+	if (history->has_records &&
+		!strncmp(ctx->meta.path, RECORDS_PREFIX, strlen(RECORDS_PREFIX))) {
+		for (struct rpchandler_history_records **record =
+				 history->facilities->records;
+			 record && *record; record++) {
+			if (!strcmp(ctx->meta.path + strlen(RECORDS_PREFIX), (*record)->name))
+				return handle_records_history(ctx, *record);
+		}
+	} else if (history->has_files &&
+		!strncmp(ctx->meta.path, FILES_PREFIX, strlen(FILES_PREFIX))) {
+		for (struct rpchandler_history_files **file = history->facilities->files;
+			 file && *file; file++) {
+			if (!strcmp(ctx->meta.path + strlen(FILES_PREFIX), (*file)->name))
+				return handle_files_history();
+		}
+	} else if (!strcmp(ctx->meta.method, "getLog") && history->has_getlog) {
+		for (const char **s = history->signals; s && *s; s++) {
+			if ((ctx->meta.path[0] != '\0') &&
+				(strncmp(*s, ctx->meta.path, strlen(ctx->meta.path)) ||
+					!((*s)[strlen(ctx->meta.path)] == '\0' ||
+						(*s)[strlen(ctx->meta.path)] == '/')))
+				continue;
+
+			struct rpchistory_getlog_request *request =
+				rpchistory_getlog_request_unpack(
+					ctx->unpack, ctx->item, rpchandler_obstack(ctx));
+			if (!rpchandler_msg_valid(ctx))
+				return RPCHANDLER_MSG_DONE;
+			if (!request)
+				rpchandler_msg_send_error(
+					ctx, RPCERR_INVALID_PARAM, "{...} expected.");
+			else
+				rpcgetlog_result(history->facilities, request, ctx, ctx->meta.path);
+
+			return RPCHANDLER_MSG_DONE;
 		}
 	}
 
@@ -159,14 +242,16 @@ static const struct rpchandler_funcs rpc_funcs = {
 };
 
 rpchandler_history_t rpchandler_history_new(
-	struct rpchandler_history_facilities *facilities) {
+	struct rpchandler_history_facilities *facilities, const char **signals) {
 	rpchandler_history_t res = malloc(sizeof *res);
 	if (!res)
 		return NULL;
 
 	res->facilities = facilities;
+	res->signals = signals;
 	res->has_records = facilities->records;
 	res->has_files = facilities->files;
+	res->has_getlog = facilities->pack_getlog;
 
 	return res;
 }